requirements.txt
colorama==0.4.6
python-dotenv==1.0.1
pymongo==4.16.0
requests==2.32.5
web3==7.14.0
py-clob-client==0.34.4
